refactor: add isFile check

+++ src/handlers/basic.js
@@ -3,12 +3,16 @@ import path from 'path';
 import { OP_FAILED } from '../model/env.js';
 import { createWriteStream } from "fs";
 import { pipeline } from 'stream/promises';
+import { checkFile } from '../utils/utils.js';
 
 export const cat = async arg => {
   try {
-    const fileOpened = await fs.open(path.resolve(arg[0]));
-    const rs = fileOpened.createReadStream();
-    rs.pipe(process.stdout);
+    const isFile = await checkFile(arg[0]);
+    if (isFile) {
+      const fileOpened = await fs.open(path.resolve(arg[0]));
+      const rs = fileOpened.createReadStream();
+      rs.pipe(process.stdout);
+    } else throw new Error();
   } catch {
     console.error(OP_FAILED);
   };
@@ -16,8 +20,7 @@ export const cat = async arg => {
 
 export const add = async arg => {
   try {
-    const pathToFile = Array.isArray(arg) ? arg[0] : arg;
-    await fs.writeFile(path.resolve(pathToFile), '');
+    await fs.writeFile(path.resolve(arg[0]), '');
   } catch {
     console.error(OP_FAILED);
   };
@@ -26,7 +29,11 @@ export const add = async arg => {
 export const rename = async args => {
   const [oldName, newName] = args;
   try {
-    await fs.rename(path.resolve(oldName), path.resolve(newName));
+    const regex = /\/|\\/g;
+    const isPath = regex.test(newName);
+    const isFile = await checkFile(oldName);
+    if (isFile && !isPath) await fs.rename(path.resolve(oldName), newName);
+    else throw new Error();
   } catch {
     console.error(OP_FAILED);
   };
@@ -35,12 +42,15 @@ export const rename = async args => {
 export const copy = async args => {
   const [oldPath, newPath] = args;
   try {
-    const openedFile = await fs.open(path.resolve(oldPath));
-    const rs = openedFile.createReadStream();
-    const basename = path.basename(oldPath);
-    const ws = createWriteStream(path.resolve(newPath, basename));
-    await pipeline(rs, ws);
-    openedFile.close();
+    const isFile = await checkFile(oldPath);
+    if (isFile) {
+      const openedFile = await fs.open(path.resolve(oldPath));
+      const rs = openedFile.createReadStream();
+      const basename = path.basename(oldPath);
+      const ws = createWriteStream(path.resolve(newPath, basename));
+      await pipeline(rs, ws);
+      openedFile.close();
+    } else throw new Error();
   } catch {
     console.error(OP_FAILED);
   };
@@ -56,7 +66,9 @@ export const move = async args => {
 
 export const remove = async arg => {
   try {
-    await fs.rm(path.resolve(arg[0]));
+    const isFile = checkFile(arg[0]);
+    if (isFile) await fs.rm(path.resolve(arg[0]));
+    else throw new Error();
   } catch {
     console.error(OP_FAILED);
   };
